requirements.txt
streamlit
PyPDF2
langchain
langchain-community
langchain-openai
langchain-core
faiss-cpu
python-dotenv
spacy
openai
anthropic